// src/app/model/teacher/Teacher.model.ts
export class Teacher {
  _id: string;
  teacherFirstName: string;
  teacherLastName: string;
  teacherEmail: string;
  userName: string;
  mobileNo: string;
  role: number;
  dob: Date;
  address: string;
  city: string;
  country: string;
  password: string;
  teacherId: string;
  state: string;
  gender: string;
  qrImg: string;
  imgSrc: string;
  status: boolean;
  constructor() { }
}